sql/migrate: move migration directory logic to its own file
