+++ sql/migrate/dir.go
@@ -5,9 +5,362 @@
 package migrate
 
 import (
+	"bufio"
+	"bytes"
+	"crypto/sha256"
+	"encoding/base64"
+	"errors"
+	"fmt"
 	"io"
+	"io/fs"
+	"io/ioutil"
+	"os"
+	"path/filepath"
 	"regexp"
+	"sort"
+	"strings"
+	"text/template"
+	"time"
 )
+
+type (
+	// Dir wraps the functionality used to interact with a migration directory.
+	Dir interface {
+		fs.FS
+		// WriteFile writes the data to the named file.
+		WriteFile(string, []byte) error
+
+		// Files returns a set of files stored in this Dir to be executed on a database.
+		Files() ([]File, error)
+	}
+
+	// Formatter wraps the Format method.
+	Formatter interface {
+		// Format formats the given Plan into one or more migration files.
+		Format(*Plan) ([]File, error)
+	}
+
+	// File represents a single migration file.
+	File interface {
+		// Name returns the name of the migration file.
+		Name() string
+		// Desc returns the description of the migration File.
+		Desc() string
+		// Version returns the version of the migration File.
+		Version() string
+		// Baseline indicates if this file is marked as a baseline migration.
+		Baseline() bool
+		// Bytes returns the read content of the file.
+		Bytes() []byte
+		// Stmts returns the set of SQL statements this file holds.
+		Stmts() ([]string, error)
+	}
+)
+
+// LocalDir implements Dir for a local migration
+// directory with default Atlas formatting.
+type LocalDir struct {
+	path string
+}
+
+var _ Dir = (*LocalDir)(nil)
+
+// NewLocalDir returns a new the Dir used by a Planner to work on the given local path.
+func NewLocalDir(path string) (*LocalDir, error) {
+	fi, err := os.Stat(path)
+	if err != nil {
+		return nil, fmt.Errorf("sql/migrate: %w", err)
+	}
+	if !fi.IsDir() {
+		return nil, fmt.Errorf("sql/migrate: %q is not a dir", path)
+	}
+	return &LocalDir{path: path}, nil
+}
+
+// Path returns the local path used for opening this dir.
+func (d *LocalDir) Path() string {
+	return d.path
+}
+
+// Open implements fs.FS.
+func (d *LocalDir) Open(name string) (fs.File, error) {
+	return os.Open(filepath.Join(d.path, name))
+}
+
+// WriteFile implements Dir.WriteFile.
+func (d *LocalDir) WriteFile(name string, b []byte) error {
+	return os.WriteFile(filepath.Join(d.path, name), b, 0644)
+}
+
+// Files implements Dir.Files. It looks for all files with .sql suffix and orders them by filename.
+func (d *LocalDir) Files() ([]File, error) {
+	names, err := fs.Glob(d, "*.sql")
+	if err != nil {
+		return nil, err
+	}
+	// Sort files lexicographically.
+	sort.Slice(names, func(i, j int) bool {
+		return names[i] < names[j]
+	})
+	ret := make([]File, len(names))
+	for i, n := range names {
+		b, err := fs.ReadFile(d, n)
+		if err != nil {
+			return nil, fmt.Errorf("sql/migrate: read file %q: %w", n, err)
+		}
+		ret[i] = &LocalFile{n: n, b: b}
+	}
+	return ret, nil
+}
+
+// LocalFile is used by LocalDir to implement the Scanner interface.
+type LocalFile struct {
+	n string
+	b []byte
+}
+
+var _ File = (*LocalFile)(nil)
+
+// NewLocalFile returns a new local file.
+func NewLocalFile(name string, data []byte) *LocalFile {
+	return &LocalFile{n: name, b: data}
+}
+
+// Name implements File.Name.
+func (f LocalFile) Name() string {
+	return f.n
+}
+
+// Baseline reports if this file marked as a baseline migration.
+func (f LocalFile) Baseline() bool {
+	_, ok := Directive(string(f.b), directivePrefixSQL, directiveBaseline)
+	return ok
+}
+
+// Desc implements File.Desc.
+func (f LocalFile) Desc() string {
+	parts := strings.SplitN(f.n, "_", 2)
+	if len(parts) == 1 {
+		return ""
+	}
+	return strings.TrimSuffix(parts[1], ".sql")
+}
+
+// Version implements File.Version.
+func (f LocalFile) Version() string {
+	return strings.SplitN(f.n, "_", 2)[0]
+}
+
+// Stmts implements Scanner.Stmts. It reads migration file line-by-line and expects a statement to be one line only.
+func (f LocalFile) Stmts() ([]string, error) {
+	return stmts(string(f.b))
+}
+
+// Bytes returns local file data.
+func (f LocalFile) Bytes() []byte {
+	return f.b
+}
+
+var (
+	// templateFuncs contains the template.FuncMap for the DefaultFormatter.
+	templateFuncs = template.FuncMap{"now": func() string { return time.Now().UTC().Format("20060102150405") }}
+	// DefaultFormatter is a default implementation for Formatter.
+	DefaultFormatter = &TemplateFormatter{
+		templates: []struct{ N, C *template.Template }{
+			{
+				N: template.Must(template.New("").Funcs(templateFuncs).Parse(
+					"{{ now }}{{ with .Name }}_{{ . }}{{ end }}.sql",
+				)),
+				C: template.Must(template.New("").Funcs(templateFuncs).Parse(
+					`{{ range .Changes }}{{ with .Comment }}-- {{ println . }}{{ end }}{{ printf "%s;\n" .Cmd }}{{ end }}`,
+				)),
+			},
+		},
+	}
+)
+
+// TemplateFormatter implements Formatter by using templates.
+type TemplateFormatter struct {
+	templates []struct{ N, C *template.Template }
+}
+
+// NewTemplateFormatter creates a new Formatter working with the given templates.
+//
+//	migrate.NewTemplateFormatter(
+//		template.Must(template.New("").Parse("{{now.Unix}}{{.Name}}.sql")),                 // name template
+//		template.Must(template.New("").Parse("{{range .Changes}}{{println .Cmd}}{{end}}")), // content template
+//	)
+//
+func NewTemplateFormatter(templates ...*template.Template) (*TemplateFormatter, error) {
+	if n := len(templates); n == 0 || n%2 == 1 {
+		return nil, fmt.Errorf("zero or odd number of templates given")
+	}
+	t := new(TemplateFormatter)
+	for i := 0; i < len(templates); i += 2 {
+		t.templates = append(t.templates, struct{ N, C *template.Template }{templates[i], templates[i+1]})
+	}
+	return t, nil
+}
+
+// Format implements the Formatter interface.
+func (t *TemplateFormatter) Format(plan *Plan) ([]File, error) {
+	files := make([]File, 0, len(t.templates))
+	for _, tpl := range t.templates {
+		var n, b bytes.Buffer
+		if err := tpl.N.Execute(&n, plan); err != nil {
+			return nil, err
+		}
+		if err := tpl.C.Execute(&b, plan); err != nil {
+			return nil, err
+		}
+		files = append(files, &LocalFile{
+			n: n.String(),
+			b: b.Bytes(),
+		})
+	}
+	return files, nil
+}
+
+// HashFileName of the migration directory integrity sum file.
+const HashFileName = "atlas.sum"
+
+// HashFile represents the integrity sum file of the migration dir.
+type HashFile []struct{ N, H string }
+
+// HashSum reads the whole dir, sorts the files by name and creates a HashSum from its contents.
+// If a files first line matches the above regex, it will be excluded from hash creation.
+func HashSum(dir Dir) (HashFile, error) {
+	var (
+		hs HashFile
+		h  = sha256.New()
+	)
+	err := fs.WalkDir(dir, "", func(path string, d fs.DirEntry, err error) error {
+		if err != nil {
+			return err
+		}
+		// If this is the integrity sum file do not include it into the sum.
+		if filepath.Base(path) == HashFileName {
+			return nil
+		}
+		if !d.IsDir() {
+			f, err := dir.Open(path)
+			if err != nil {
+				return err
+			}
+			defer f.Close()
+			if _, err := h.Write([]byte(path)); err != nil {
+				return err
+			}
+			c, err := ioutil.ReadAll(f)
+			if err != nil {
+				return err
+			}
+			// Check if this file contains an "atlas:sum" directive and if so, act to it.
+			if mode, ok := Directive(string(c), "", directiveSum); ok && mode == sumModeIgnore {
+				return nil
+			}
+			if _, err := h.Write(c); err != nil {
+				return err
+			}
+			hs = append(hs, struct{ N, H string }{path, base64.StdEncoding.EncodeToString(h.Sum(nil))})
+		}
+		return nil
+	})
+	if err != nil {
+		return nil, err
+	}
+	return hs, nil
+}
+
+// WriteSumFile writes the given HashFile to the Dir. If the file does not exist, it is created.
+func WriteSumFile(dir Dir, sum HashFile) error {
+	b, err := sum.MarshalText()
+	if err != nil {
+		return err
+	}
+	return dir.WriteFile(HashFileName, b)
+}
+
+// Sum returns the checksum of the represented hash file.
+func (f HashFile) Sum() string {
+	sha := sha256.New()
+	for _, f := range f {
+		sha.Write([]byte(f.N))
+		sha.Write([]byte(f.H))
+	}
+	return base64.StdEncoding.EncodeToString(sha.Sum(nil))
+}
+
+// MarshalText implements encoding.TextMarshaler.
+func (f HashFile) MarshalText() ([]byte, error) {
+	buf := new(bytes.Buffer)
+	for _, f := range f {
+		fmt.Fprintf(buf, "%s h1:%s\n", f.N, f.H)
+	}
+	return []byte(fmt.Sprintf("h1:%s\n%s", f.Sum(), buf)), nil
+}
+
+// UnmarshalText implements encoding.TextUnmarshaler.
+func (f *HashFile) UnmarshalText(b []byte) error {
+	sc := bufio.NewScanner(bytes.NewReader(b))
+	// The first line contains the sum.
+	sc.Scan()
+	sum := strings.TrimPrefix(sc.Text(), "h1:")
+	for sc.Scan() {
+		li := strings.SplitN(sc.Text(), "h1:", 2)
+		if len(li) != 2 {
+			return ErrChecksumFormat
+		}
+		*f = append(*f, struct{ N, H string }{strings.TrimSpace(li[0]), li[1]})
+	}
+	if sum != f.Sum() {
+		return ErrChecksumMismatch
+	}
+	return sc.Err()
+}
+
+func (f HashFile) sumByName(n string) (string, error) {
+	for _, f := range f {
+		if f.N == n {
+			return f.H, nil
+		}
+	}
+	return "", errors.New("checksum not found")
+}
+
+var (
+	// ErrChecksumFormat is returned from Validate if the sum files format is invalid.
+	ErrChecksumFormat = errors.New("checksum file format invalid")
+	// ErrChecksumMismatch is returned from Validate if the hash sums don't match.
+	ErrChecksumMismatch = errors.New("checksum mismatch")
+	// ErrChecksumNotFound is returned from Validate if the hash file does not exist.
+	ErrChecksumNotFound = errors.New("checksum file not found")
+)
+
+// Validate checks if the migration dir is in sync with its sum file.
+// If they don't match ErrChecksumMismatch is returned.
+func Validate(dir Dir) error {
+	fh, err := readHashFile(dir)
+	if errors.Is(err, fs.ErrNotExist) {
+		// If there are no migration files yet this is okay.
+		files, err := fs.ReadDir(dir, "/")
+		if err != nil || len(files) > 0 {
+			return ErrChecksumNotFound
+		}
+		return nil
+	}
+	if err != nil {
+		return err
+	}
+	mh, err := HashSum(dir)
+	if err != nil {
+		return err
+	}
+	if fh.Sum() != mh.Sum() {
+		return ErrChecksumMismatch
+	}
+	return nil
+}
 
 // SplitBaseline splits the directory files into two groups.
 // The first one holds all baseline files and the second
